+++ zhmcclient/_hba.py
@@ -57,6 +57,7 @@ class HbaManager(BaseManager):
 
         super(HbaManager, self).__init__(
             resource_class=Hba,
+            session=partition.manager.session,
             parent=partition,
             uri_prop='element-uri',
             name_prop='name',
@@ -120,6 +121,8 @@ class HbaManager(BaseManager):
                     resource_obj_list.append(resource_obj)
                     if full_properties:
                         resource_obj.pull_full_properties()
+
+        self._name_uri_cache.update_from(resource_obj_list)
         return resource_obj_list
 
     @logged_api_call
@@ -163,7 +166,11 @@ class HbaManager(BaseManager):
         # returned props should overwrite the input props:
         props = properties.copy()
         props.update(result)
-        return Hba(self, props['element-uri'], None, props)
+        name = props.get(self._name_prop, None)
+        uri = props[self._uri_prop]
+        hba = Hba(self, uri, name, props)
+        self._name_uri_cache.update(name, uri)
+        return hba
 
 
 class Hba(BaseResource):
@@ -217,6 +224,8 @@ class Hba(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         self.manager.session.delete(self._uri)
+        self.manager._name_uri_cache.delete(
+            self.properties.get(self.manager._name_prop, None))
 
     @logged_api_call
     def update_properties(self, properties):
@@ -245,7 +254,10 @@ class Hba(BaseResource):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        self.manager.session.post(self._uri, body=properties)
+        self.manager.session.post(self.uri, body=properties)
+        self.properties.update(properties.copy())
+        if self.manager._name_prop in properties:
+            self.manager._name_uri_cache.update(self.name, self.uri)
 
     @logged_api_call
     def reassign_port(self, port):