+++ zhmcclient/_virtual_function.py
@@ -56,6 +56,7 @@ class VirtualFunctionManager(BaseManager):
         #     Partition defining the scope for this manager.
         super(VirtualFunctionManager, self).__init__(
             resource_class=VirtualFunction,
+            session=partition.manager.session,
             parent=partition,
             uri_prop='element-uri',
             name_prop='name',
@@ -119,6 +120,8 @@ class VirtualFunctionManager(BaseManager):
                     resource_obj_list.append(resource_obj)
                     if full_properties:
                         resource_obj.pull_full_properties()
+
+        self._name_uri_cache.update_from(resource_obj_list)
         return resource_obj_list
 
     @logged_api_call
@@ -158,7 +161,11 @@ class VirtualFunctionManager(BaseManager):
         # returned props should overwrite the input props:
         props = properties.copy()
         props.update(result)
-        return VirtualFunction(self, props['element-uri'], None, props)
+        name = props.get(self._name_prop, None)
+        uri = props[self._uri_prop]
+        vf = VirtualFunction(self, uri, name, props)
+        self._name_uri_cache.update(name, uri)
+        return vf
 
 
 class VirtualFunction(BaseResource):
@@ -212,6 +219,8 @@ class VirtualFunction(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         self.manager.session.delete(self._uri)
+        self.manager._name_uri_cache.delete(
+            self.properties.get(self.manager._name_prop, None))
 
     @logged_api_call
     def update_properties(self, properties):
@@ -240,4 +249,7 @@ class VirtualFunction(BaseResource):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        self.manager.session.post(self._uri, body=properties)
+        self.manager.session.post(self.uri, body=properties)
+        self.properties.update(properties.copy())
+        if self.manager._name_prop in properties:
+            self.manager._name_uri_cache.update(self.name, self.uri)