+++ zhmcclient/_virtual_switch.py
@@ -69,6 +69,7 @@ class VirtualSwitchManager(BaseManager):
 
         super(VirtualSwitchManager, self).__init__(
             resource_class=VirtualSwitch,
+            session=cpc.manager.session,
             parent=cpc,
             uri_prop='object-uri',
             name_prop='name',
@@ -141,6 +142,7 @@ class VirtualSwitchManager(BaseManager):
                     if full_properties:
                         resource_obj.pull_full_properties()
 
+        self._name_uri_cache.update_from(resource_obj_list)
         return resource_obj_list
 
 
@@ -254,3 +256,6 @@ class VirtualSwitch(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         self.manager.session.post(self.uri, body=properties)
+        self.properties.update(properties.copy())
+        if self.manager._name_prop in properties:
+            self.manager._name_uri_cache.update(self.name, self.uri)